+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.8] - 2026-02-14
+
+### Added
+
+- uncontrolled and controlled state management to `TabSwitcher`
+- `initiallyActive` attribute to `TabPanel`s
+
 ## [0.8.7] - 2026-02-13
 
 ### Changed